Some Sound Fixes, some preliminary Rendering Stuff, some preliminary Preferences Stuff
darcs-hash:20051021143405-7aa3a-d9057e343a7684634f926b8e6d9eef8a4c979618

+++ timeline/TimelineBase.cxx
@@ -124,7 +124,7 @@ Clip* TimelineBase::find( int track, int64_t position )
 static int track_length_helper( void* p, void* data )
 {
 	int64_t l;
-	int* max = (int*)data;
+	int64_t* max = (int64_t*)data;
 	track_node* node = (track_node*)p;
 	l = node->track->length();
 	if ( l > *max ) {
